037 - Stripped everything from dashboard and made pinned task card component
