+++ task-management-system/imports/ui/components/pages/dashboard/PinnedTasks.css
@@ -1,4 +1,4 @@
-/* Dashboard CSS */
+/* Pinned tasks CSS */
 
 .dashboard-pinned-container {
     max-width: calc(100vw - 300px);
